requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.2.0
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.1
django-filter==23.5
gunicorn==21.2.0
whitenoise==6.6.0
pytest==8.0.0
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.6.0
drf-yasg==1.21.7 